+++ src/main/java/br/unisul/revendaunisul/view/TelaListagemColaboradores.java
@@ -1,5 +1,7 @@
 package br.unisul.revendaunisul.view;
 
+import java.awt.event.WindowAdapter;
+import java.awt.event.WindowEvent;
 import java.util.List;
 
 import javax.swing.GroupLayout;
@@ -17,6 +19,7 @@ import javax.swing.border.EmptyBorder;
 import javax.swing.table.TableColumnModel;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
 
 import br.unisul.revendaunisul.entity.Colaborador;
@@ -37,11 +40,27 @@ public class TelaListagemColaboradores extends JFrame {
 	@Autowired
 	private TelaCadastroColaboradores cadastro;
 	
+	@Autowired
+	@Lazy
+	private TelaPrincipal telaPrincipal;
+	
+	public void atualizarTabela() {
+		table.updateUI();
+	}
+	
 	public TelaListagemColaboradores() {
 		setTitle("Gerenciar Colaboradores");
 		setDefaultCloseOperation(JFrame.HIDE_ON_CLOSE);
 		setBounds(100, 100, 450, 300);
 		setLocationRelativeTo(null);
+		
+		//captura o evento de click no X e redireciona para a tela principal
+		addWindowListener(new WindowAdapter() {
+			public void windowClosing(WindowEvent e) {
+				setVisible(false);
+				telaPrincipal.setVisible(true);
+			}
+		});
 		
 		contentPane = new JPanel();
 		contentPane.setBorder(new EmptyBorder(5, 5, 5, 5));
@@ -62,7 +81,7 @@ public class TelaListagemColaboradores extends JFrame {
 			cm.getColumn(0).setPreferredWidth(75);
 			cm.getColumn(2).setPreferredWidth(80);
 			cm.getColumn(3).setPreferredWidth(100);
-			table.updateUI();
+			atualizarTabela();
 		});
 		
 		JButton btnNovo = new JButton("Novo");
@@ -75,10 +94,17 @@ public class TelaListagemColaboradores extends JFrame {
 			int linhaSelecionada = table.getSelectedRow();
 			ColaboradorTableModel model = (ColaboradorTableModel) table.getModel();
 			Colaborador colaboradorSalvo = model.getBy(linhaSelecionada);
-			service.excluirPor(colaboradorSalvo.getId());
-			model.removeBy(linhaSelecionada);
-			table.updateUI();
-			JOptionPane.showMessageDialog(contentPane, "Colaborador removido com sucesso!");
+			
+			int opcaoSelecionada = JOptionPane.showConfirmDialog(
+					contentPane, "Deseja realmente remover?", "Confirmação", JOptionPane.YES_NO_OPTION);
+			
+			if(opcaoSelecionada == JOptionPane.YES_OPTION) {
+				service.excluirPor(colaboradorSalvo.getId());
+				model.removeBy(linhaSelecionada);
+				JOptionPane.showMessageDialog(contentPane, "Colaborador removido com sucesso!");
+			}
+			
+			atualizarTabela();
 		});
 		
 		JButton btnEditar = new JButton("Editar");
@@ -97,8 +123,8 @@ public class TelaListagemColaboradores extends JFrame {
 		
 		GroupLayout gl_contentPane = new GroupLayout(contentPane);
 		gl_contentPane.setHorizontalGroup(
-			gl_contentPane.createParallelGroup(Alignment.LEADING)
-				.addGroup(Alignment.TRAILING, gl_contentPane.createSequentialGroup()
+			gl_contentPane.createParallelGroup(Alignment.TRAILING)
+				.addGroup(gl_contentPane.createSequentialGroup()
 					.addContainerGap()
 					.addGroup(gl_contentPane.createParallelGroup(Alignment.TRAILING)
 						.addComponent(scrollPane, GroupLayout.DEFAULT_SIZE, 404, Short.MAX_VALUE)
@@ -109,11 +135,11 @@ public class TelaListagemColaboradores extends JFrame {
 							.addPreferredGap(ComponentPlacement.RELATED)
 							.addComponent(btnBuscar))
 						.addGroup(gl_contentPane.createSequentialGroup()
-							.addComponent(btnEditar, GroupLayout.PREFERRED_SIZE, 65, GroupLayout.PREFERRED_SIZE)
+							.addComponent(btnEditar, GroupLayout.PREFERRED_SIZE, 77, GroupLayout.PREFERRED_SIZE)
 							.addPreferredGap(ComponentPlacement.RELATED)
-							.addComponent(btnExcluir, GroupLayout.PREFERRED_SIZE, 65, GroupLayout.PREFERRED_SIZE)
+							.addComponent(btnExcluir, GroupLayout.PREFERRED_SIZE, 75, GroupLayout.PREFERRED_SIZE)
 							.addPreferredGap(ComponentPlacement.RELATED)
-							.addComponent(btnNovo, GroupLayout.PREFERRED_SIZE, 65, GroupLayout.PREFERRED_SIZE)))
+							.addComponent(btnNovo, GroupLayout.PREFERRED_SIZE, 74, GroupLayout.PREFERRED_SIZE)))
 					.addContainerGap())
 		);
 		gl_contentPane.setVerticalGroup(
@@ -135,4 +161,5 @@ public class TelaListagemColaboradores extends JFrame {
 		);
 		contentPane.setLayout(gl_contentPane);
 	}
+
 }
